Bug fix on interpolation
Interpolation did not work when messages arrived unordered.

Now interpolation works only when the last three more recent messages are close
enough and the second one is the only one possibly interpolating.

+++ src/backend/Persistence.MongoDB/Servizi/StoreMessaggioPosizione_DeleteInterpolatedMessages_Decorator.cs
@@ -47,6 +47,13 @@ namespace Persistence.MongoDB.Servizi
         /// </summary>
         public float InterpolationThreshold_mt { get; set; }
 
+        /// <summary>
+        ///   Try to interpolate the three most recent messages. Interpolation only works if the
+        ///   first and the third message do not interpolate.
+        /// </summary>
+        /// <param name="newMessage">
+        ///   The message to be stored. Storage is in charge of the decorated instance.
+        /// </param>
         public void Store(MessaggioPosizione newMessage)
         {
             this.decorated.Store(newMessage);
@@ -57,30 +64,47 @@ namespace Persistence.MongoDB.Servizi
                 .Limit(3)
                 .ToList();
 
-            if (lastThreeMessages.Count == 3)
+            if ((lastThreeMessages.Count == 3) &&
+                firstAndThirdMessagesDoNotInterpolate(lastThreeMessages) &&
+                this.AllInTheSamePosition(lastThreeMessages))
             {
-                if (this.AllInTheSamePosition(lastThreeMessages))
-                {
-                    var interpolatingMessage = lastThreeMessages[0];
-                    var msgToInterpolate = lastThreeMessages[1];
-                    var lastInterpolationData = msgToInterpolate.InterpolationData ?? new InterpolationData(0, 0, null);
+                var interpolatingMessage = lastThreeMessages[0];
+                var msgToInterpolate = lastThreeMessages[1];
+                var lastInterpolationData = msgToInterpolate.InterpolationData ?? new InterpolationData(0, 0, null);
 
-                    var interpolationData = new InterpolationData(
-                        (int)(lastInterpolationData.Length_sec +
-                            newMessage.IstanteAcquisizione.Subtract(msgToInterpolate.IstanteAcquisizione).TotalSeconds),
-                        lastInterpolationData.Messages + 1,
-                        msgToInterpolate.IstanteAcquisizione);
+                var interpolationData = new InterpolationData(
+                    (int)(lastInterpolationData.Length_sec +
+                        newMessage.IstanteAcquisizione.Subtract(msgToInterpolate.IstanteAcquisizione).TotalSeconds),
+                    lastInterpolationData.Messages + 1,
+                    msgToInterpolate.IstanteAcquisizione);
 
-                    var deleteTask = this.messaggiPosizioneCollection.DeleteOneAsync(m => m.Id == msgToInterpolate.Id);
-                    var updateTask = this.messaggiPosizioneCollection.UpdateOneAsync(
-                        m => m.Id == interpolatingMessage.Id,
-                        Builders<MessaggioPosizione>.Update.Set(m => m.InterpolationData, interpolationData));
+                var deleteTask = this.messaggiPosizioneCollection.DeleteOneAsync(m => m.Id == msgToInterpolate.Id);
+                var updateTask = this.messaggiPosizioneCollection.UpdateOneAsync(
+                    m => m.Id == interpolatingMessage.Id,
+                    Builders<MessaggioPosizione>.Update.Set(m => m.InterpolationData, interpolationData));
 
-                    Task.WaitAll(deleteTask, updateTask);
-                }
+                Task.WaitAll(deleteTask, updateTask);
             }
         }
 
+        /// <summary>
+        ///   Checks whether the list contains three messages and the first and the third do not interpolate.
+        /// </summary>
+        /// <param name="lastThreeMessages">The list to be checked</param>
+        /// <returns>Returns true if and only if the check passes</returns>
+        private bool firstAndThirdMessagesDoNotInterpolate(List<MessaggioPosizione> lastThreeMessages)
+        {
+            return
+                (lastThreeMessages.Count == 3) &&
+                (lastThreeMessages[0].InterpolationData == null) &&
+                (lastThreeMessages[2].InterpolationData == null);
+        }
+
+        /// <summary>
+        ///   Checks whether the list contains messages falling close one each other.
+        /// </summary>
+        /// <param name="messages">The messages to be checked</param>
+        /// <returns>True if (and only if) the message are close enough</returns>
         private bool AllInTheSamePosition(IList<MessaggioPosizione> messages)
         {
             var first = messages.First();
@@ -89,6 +113,11 @@ namespace Persistence.MongoDB.Servizi
             return allButFirst.All(m => this.AreCloseEnough(first.Localizzazione, m.Localizzazione));
         }
 
+        /// <summary>
+        ///   Checks whether two messages fall close one another.
+        /// </summary>
+        /// <param name="messages">The messages to be checked</param>
+        /// <returns>True if (and only if) the message are close enough</returns>
         private bool AreCloseEnough(Localizzazione loc1, Localizzazione loc2)
         {
             var distance = loc1.GetDistanceTo(loc2);
